made the ExamSeat a JSON field in ExamRoom, rewrote a bunch of stuff to make it work again, moved the upload button to the top of the page,liquibase changes, removed the ExamSeat foreign key from ExamUser

+++ src/main/java/de/tum/cit/aet/artemis/exam/dto/room/ExamRoomDTO.java
@@ -10,7 +10,7 @@ public record ExamRoomDTO(
     String roomNumber,
     String name,
     String building,
-    Long numberOfSeats,
+    Integer numberOfSeats,
     Set<ExamRoomLayoutStrategyDTO> layoutStrategies
 ) {
 }
